# sync-server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fractonica API
  version: current
  summary: Server-authoritative journal, event, and synchronization API.
  description: |
    The Fractonica API is the canonical server contract.
    It is multi-user, cursor-synchronized, and intended for the first-party iOS
    client as well as user-authorized automation agents.

    Authentication uses either a short-lived JWT access token (`JwtBearer`) or
    a user-issued API key (`ApiKeyBearer`) in the `Authorization: Bearer ...`
    header. API keys are bound to one user and one logical device. Operations
    carrying `x-required-scopes` enforce those scopes for API-key callers; JWT
    sessions act with the authenticated user's full authority.

    Production deployments must expose the API only over HTTPS/TLS and reject
    plain HTTP at the edge. Passwords and bearer credentials are otherwise
    observable in transit. Plain HTTP is for loopback development only.

    Records are public by default. Public record payloads are readable and may
    be rendered from a template. Private record payloads and private
    media are encrypted by the client before upload; the server stores only an
    authenticated ciphertext envelope and intentionally cannot render, index,
    or embed their plaintext.

    All timestamps are RFC 3339 instants. Record API identifiers are globally
    unique five-character Base64URL strings; their owner-only `originId` UUIDs
...
    are UUIDs.
...

...
    immediately retrying.

...
  license:
...
servers:
...
security:
  - JwtBearer: []
  - ApiKeyBearer: []
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [Events]
      summary: List lightweight user events
      description: |
        Events are owner-scoped and ordered by `startsAt` descending and ID.
        When `from` or `to` is supplied, events are selected by interval overlap.
      x-required-scopes: [events:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/DeviceIdFilter'
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, private]
        - name: type
          in: query
          description: Repeat to include more than one numeric event type.
          style: form
          explode: true
          schema:
            type: array
            maxItems: 50
            items:
              type: integer
              format: int32
              minimum: 0
        - name: from
          in: query
          description: Inclusive lower bound for interval overlap.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Exclusive upper bound for interval overlap.
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/UpdatedAfter'
      responses:
        '200':
          description: A stable cursor page of events.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      operationId: createEvent
      tags: [Events]
      summary: Create a lightweight event
      description: |
        Events have no media, templates, encryption envelope, or embeddings.
        They are public by default; visibility is immutable after creation.
        `userId` is always derived from the credential. API-key callers must use
        the key's bound device ID. Omitted `endsAt` means ongoing; an `endsAt`
        equal to `startsAt` represents an instantaneous event.
      x-required-scopes: [events:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
            example:
              id: 20a78723-c33e-4794-a036-5da69a15e8bf
              deviceId: 2dca8eab-00a8-4e94-9bd2-2fcbfe17e890
              startsAt: '2026-07-14T16:42:00Z'
              endsAt: '2026-07-14T16:55:00Z'
              label: X1.7 solar flare
              type: 1001
              metadata:
                source: noaa-swpc
                activeRegion: 4136
      responses:
        '201':
          description: Event created.
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /events/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      operationId: getEvent
      tags: [Events]
      summary: Get a lightweight event
      x-required-scopes: [events:read]
      responses:
        '200':
          description: Current event revision.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    patch:
      operationId: updateEvent
      tags: [Events]
      summary: Patch a lightweight event
      x-required-scopes: [events:write]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated and revision incremented.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      operationId: deleteEvent
      tags: [Events]
      summary: Soft-delete a lightweight event
      x-required-scopes: [events:write]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Event deleted and tombstone emitted.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /public/events:
    get:
      operationId: listPublicEvents
      tags: [Events]
      summary: List public events
      description: Public events are ordered by occurrence time and never expose a device ID.
      security: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          style: form
          explode: true
          schema:
            type: array
            maxItems: 50
            items:
              type: integer
              format: int32
              minimum: 0
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Public event page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicEventPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /public/events/{eventId}:
    get:
      operationId: getPublicEvent
      tags: [Events]
      summary: Get one public event
      security: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Public event projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicEventProjection'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    JwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived Ed25519-signed access JWT issued by login or refresh. The
        token includes `iss`, `aud`, `sub`, `exp`, and `kid` claims.
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: Fractonica API key
      description: |
        User-issued secret beginning with `exk_`. Keys are stored hashed,
        device-bound, scoped, independently expirable, and revocable. The
        plaintext secret appears only in the initial creation response; a replay
        returns the created `apiKeyId`, never the secret.
  parameters:
...
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the immediately preceding page.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-generated identifier for one logical mutation. It must be reused
        for retries of the same body and must not be reused for another body.
        Most operations replay their original response. API-key creation is the
        deliberate exception: replay returns a 409 problem containing the
        created `apiKeyId` because its plaintext secret is returned only once.
      schema:
        type: string
        minLength: 8
        maxLength: 255
        pattern: '^[\x21-\x7E]+$'
      example: 01J2V7YCB9Q3WZ2P6AQ03M6DN8
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag of the revision being mutated.
      schema:
        type: string
        minLength: 3
        maxLength: 200
      example: '"record-aB_9Z-r4"'
...
    DeviceIdFilter:
      name: deviceId
      in: query
      schema:
        type: string
        format: uuid
...
    UpdatedAfter:
      name: updatedAfter
      in: query
      description: Inclusive lower bound on server update time.
      schema:
        type: string
        format: date-time
...
  headers:
    ETag:
      description: Strong opaque validator for the returned resource revision.
      schema:
        type: string
      example: '"record-aB_9Z-r4"'
    RequestId:
      description: Server-assigned request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: The request syntax, parameter combination, or cursor is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.exeligmos.app/problems/invalid-request
            title: Invalid request
            status: 400
            detail: The occurredAfter timestamp must precede occurredBefore.
            code: invalid_request
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    Unauthorized:
      description: Authentication is missing, expired, revoked, or invalid.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="exeligmos"
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the scope or ownership required for this operation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The resource is absent or is not visible to this caller.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: The mutation conflicts with current state or idempotency history.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    PreconditionFailed:
      description: "`If-Match` does not match the current resource revision."
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    UnprocessableContent:
      description: The request is syntactically valid but violates domain constraints.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    ServiceUnavailable:
      description: A database statement or lock exceeded its safety deadline; the operation may be retried.
      headers:
        Retry-After:
          description: Minimum number of seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
          example: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:exeligmos:problem:database-timeout
            title: Service Unavailable
            status: 503
            detail: The database could not complete the request within its safety deadline.
            code: database_timeout
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    TooManyRequests:
      description: Rate or concurrency limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details with stable application extensions.
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem type URI.
        title:
          type: string
          minLength: 1
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        requestId:
          type: string
          minLength: 1
          maxLength: 128
      additionalProperties: true
...
    ResourceMetadata:
      type: object
      description: |
        Must serialize to at most 32768 UTF-8 bytes. Object keys and string
        values must not contain U+0000 or unpaired UTF-16 surrogates.
      x-max-serialized-bytes: 32768
      additionalProperties: true
      default: {}
...
    CursorPageMeta:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          type: string
          minLength: 1
          maxLength: 2048
        hasMore:
          type: boolean
...
    CreateEventRequest:
      type: object
      required: [deviceId, startsAt, label, type]
      properties:
        id:
          type: string
          format: uuid
          description: Optional stable client-generated UUID.
        deviceId:
          type: string
          format: uuid
        visibility:
          type: string
          enum: [public, private]
          default: public
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
          description: Must be equal to or later than `startsAt`.
        label:
          type: string
          minLength: 1
          maxLength: 256
        type:
          type: integer
          format: int32
          minimum: 0
          description: |
            Stable numeric event type. Clients must preserve and tolerate
            unknown future values rather than rejecting the event.
        metadata:
          $ref: '#/components/schemas/ResourceMetadata'
        references:
          $ref: '#/components/schemas/ResourceReferenceList'
      additionalProperties: false
    UpdateEventRequest:
      type: object
      minProperties: 1
      properties:
        deviceId:
          type: string
          format: uuid
        visibility:
          type: string
          enum: [public, private]
          description: Immutable after creation.
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: [string, 'null']
          format: date-time
          description: Set null to mark the event as ongoing.
        label:
          type: string
          minLength: 1
          maxLength: 256
        type:
          type: integer
          format: int32
          minimum: 0
        metadata:
          $ref: '#/components/schemas/ResourceMetadata'
        references:
          $ref: '#/components/schemas/ResourceReferenceList'
      additionalProperties: false
    Event:
      type: object
      required:
        - id
        - userId
        - author
        - deviceId
        - visibility
        - startsAt
        - label
        - type
        - metadata
        - references
        - revision
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          readOnly: true
        author:
          $ref: '#/components/schemas/PublicUserSummary'
        deviceId:
          type: string
          format: uuid
        visibility:
          type: string
          enum: [public, private]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        label:
          type: string
          minLength: 1
          maxLength: 256
        type:
          type: integer
          format: int32
          minimum: 0
        metadata:
          $ref: '#/components/schemas/ResourceMetadata'
        references:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/ResourceReference'
        revision:
          type: integer
          format: int64
          minimum: 1
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      additionalProperties: false
    EventPage:
      allOf:
        - $ref: '#/components/schemas/CursorPageMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Event'
    PublicEventProjection:
      type: object
      required:
        [id, userId, author, visibility, startsAt, label, type, metadata, references, revision, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/PublicUserSummary'
        visibility:
          type: string
          const: public
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        label:
          type: string
          minLength: 1
          maxLength: 256
        type:
          type: integer
          format: int32
          minimum: 0
        metadata:
          $ref: '#/components/schemas/ResourceMetadata'
        references:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/ResourceReference'
        revision:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    PublicEventPage:
      allOf:
        - $ref: '#/components/schemas/CursorPageMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PublicEventProjection'
...